requirements.txt
fredapi>=0.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
scipy>=1.11.0
mlflow>=2.7.0
joblib>=1.3.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
